Module Camlkit.Appkit_AppDelegate

module type S = sig ... end

module Create (D : S) : sig ... end

Note: get_protocol "NSApplicationDelegate" fails since its object is not created by the runtime unless referenced in ObjC code: https://stackoverflow.com/questions/10212119/objc-getprotocol-returns-null-for-nsapplicationdelegate But it's an informal protocol, not required for the code to function.