Module Appkit_AppDelegate.Create

Note: get_protocol "NSApplicationDelegate" fails since its object is not created by the runtime unless referenced in ObjC code: https://stackoverflow.com/questions/10212119/objc-getprotocol-returns-null-for-nsapplicationdelegate But it's an informal protocol, not required for the code to function.

Parameters

module D : S

Signature

val self : Runtime__.Type_description.Types.objc_object Ctypes.structure Ctypes_static.ptr