Parameter Create.D

val class_name : string

val on_before_start : Runtime.Objc.object_t -> unit

Tells the delegate that the app's initialization is about to complete.

val on_started : Runtime.Objc.object_t -> unit

Tells the delegate that the launch process is almost done and the app is almost ready to run.

val on_before_terminate : Runtime.Objc.object_t -> unit

Tells the delegate that the app is about to terminate.

val terminate_on_windows_closed : Runtime.Objc.object_t -> bool

Returns a Boolean value that indicates if the app terminates once the last window closes.