Module Camlkit.Appkit_global

module StyleMask : sig ... end

module BackingStoreType : sig ... end

module ActivationPolicy : sig ... end

module ActivationOptions : sig ... end

val main : argc:int -> argv:string Ctypes_static.ptr -> int

Called by the main function to create and run the application.

val shared_application : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr

val shared : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr

val set_main_menu : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

val set_activation_policy : Signed.llong -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> bool

val activate_ignoring_other_apps : bool -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

val run : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

val set_delegate : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

val set_title : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

val init_with_frame : [ `CGRect ] Runtime.Objc.structure -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr

val set_target : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

val set_action : Runtime.Objc.objc_selector Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

val content_view : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr

val add_subview : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

val set_frame : [ `CGRect ] Runtime.Objc.structure -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

val init_with_title : string -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr

Initializes and returns a menu having the specified title and with autoenabling of menu items turned on.

val set_submenu : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> for_item:Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

Assigns a menu to be a submenu of the menu controlled by a given menu item.

val add_item : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> unit

Adds a menu item to the end of the menu.

val add_item' : Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr -> title:string -> action:Runtime.Objc.objc_selector Ctypes.structure Ctypes_static.ptr -> key_equivalent:string -> Runtime.Objc.objc_object Ctypes.structure Ctypes_static.ptr

Creates a new menu item and adds it to the end of the menu.

module Button : sig ... end

module Label : sig ... end

module View : sig ... end